Restructure the currency model
It proves to be unfeasible for now to use a model embedded in an embeddable
field like money. Therefore the currency is now a simple wrapper around
the iso currency code. This provides some restrictions on what can go
into a money as a currency and leaves room to later reexplore the
possibility of storing currency properties in the database using a
currency model