Changed money, small bug with Cost not having it's own equals function
Wrote an automated consistency checker for stock, and two tests for it. One test is quick, and shows it works.
The other is disabled by default and takes around 5 minutes. It creates an extremely simple test database, adds a single error (among 100 000 lines of StockChanges) and checks the results.
The data generation takes 280 seconds, testing takes 3.9