PEP8-ed the code.
Added a small rework of the transaction code. Added explanations for all classes. Added proper inheritance for all transaction lines. Added payment type to payment and made sure that if a transaction is created, all checks are done before _any_ amount of saving is done.

Modded tests to comply to new scenario.