Added transitions to OrderLines and added tests
OrderLines are able to change from one state to another. Functions are used to ensure data is not corrupted when OrderLines transition from one state to another. Furthermore, illegal states will not be accepted when using any means to store the OrderLine or when transitioning its state.

OrderLines keep a log of their transitions in the combined OrderLineStates. The current state of the OrderLine can be inferred from the history of transitions.