Further added information on models
Orderlines can now be added by using a method that abstracts away most complexity. Orders can now be saved together with their respective orders as this is combined in one function that handles the logic. Saved orders now have a print function that gives a quick view on the state of the orders.