Previous Orders did not account for Wishabletypes other than Articletypes. Some changes were made to change that.
Support for wishables in Orders is not complete. Since the way Wishables are handled and processed is different for each type, careful thinking about the consequences is important. As a stop gap measure, I have chosen to only implement support for Sellabletypes as vat for And or Or products can be very complicated. The system can be expanded if support in the article branch is improved.