Fix for bug in orders
A bug in orders checked if a wishable was a SellableType. This was however done incorrectly. Due to Django-internals, saved objects are upcast to their model type, which in this case was WishableType.

The hasattr function does indeed check that the object is a SellableType(and therefore mostly an articletype). This error surfaced when retrieving orderLines.