Added state to SupplierOrderLine with logging. Small improvements.
SupplierOrderLine now has a state. This means that tracking the actual working of a SupplierOrder is now a possibility. Also added logging for the state of every supplierOrder. This makes getting the relevant orders a lot easier than any potential backwards lookups from other tables. Fixed a bug where an exception is returned in stead of raised. Updated __str__ functions where applicable.