Made beginning of automated distribution function.
Added a SupplierOrderCombinationLine that, like OrderCombinationLine, combines lines that are 'similar'. It might not serve the needs of every user and therefore should be expanded until it does, but it can greatly reduce the load of the user who iterates over supplierorderlines looking for specific information. The function will be used in the automated distribution function for the checks.