Just RMAs making my life that bit harder.
Added some interaction with refunds. Refunds can now create refunds from their respective saves. This models the way the customers are refunded, namely directly via an upfront refund or indirectly if they are refunded something which is already in the RMA process. The save function contains guards to prevent collisions of options which would be impossible if both true.