Fixed bug where MoneyInOut were counted the wrong way.
MoneyInOut had an incorrect plus sign which made difference calculations fail. Fixed that and added test to catch this specific issue. More work should be done. Renamed function to functionality. Added really handy representation for Money so that it can actually be debugged from the command line instead of a debugging tool or a manual print.