tools: A lot of progress on PDF generation. Receipt, internalization and invoice templates are implemented.
I ditched reportlab for the PDF generation, because the code was really clunky and the library is not really documented.
I switched to wkhtmltopdf, which is a wrapper around the cli-tool.
This tool generates PDF's from HTML templates, so we can use the Django HTML rendering engine, which is nice.

Currently, receipt, internalization and invoice templates are implemented.

For a demo see http://localhost:8000/tools/pdftest/receipt/ (in the Django runserver, not the JS one)